DEV-6 Split `login.php` into several files and grouped them in separate folders.
